accubits/books/urls.py
from django.urls import include, path
from rest_framework import routers
from . import views

router = routers.DefaultRouter()
router.register(r'register', views.Library_usersViewSet)

router.register(r'books_borrowed', views.Book_borrowersViewSet)

router.register(r'Library_books', views.Library_booksViewSet)


# Wire up our API using automatic URL routing.
# Additionally, we include login URLs for the browsable API.
urlpatterns = [
    path('', include(router.urls)),
    path('api-auth/', include('rest_framework.urls', namespace='rest_framework'))

]