requirements.txt
asgiref==3.2.10
automium==0.2.6
automium-web==0.1.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
crypto==1.4.1
Django==3.1
django-enumfield==2.0.2
django-enumfields==2.0.0
djangorestframework==3.11.1
Flask==1.1.2
Flask-MySQL==1.5.1
flask-mysql-connector==1.0.3
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
legacy==0.1.6
login==0.0.6
MarkupSafe==1.1.1
mysql-connector-python==8.0.21
Naked==0.1.31
numpy==1.19.2
pandas==1.1.2
protobuf==3.13.0
pycryptodome==3.9.8
PyMySQL==0.10.0
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
shellescape==3.8.1
six==1.15.0
sqlparse==0.3.1
stripe==2.50.0
urllib3==1.25.10
Werkzeug==1.0.1
